+++ packages/client/src/app/components/core/Form/Form.module.scss
@@ -1,0 +1,17 @@
+@import "src/assets/styles/variables";
+
+.layout {
+  width: 100%;
+  height: 100%;
+  background: $layout-bg-blue;
+  top: 0;
+  -webkit-transition: all 0.4s;
+  -o-transition: all 0.4s;
+  -moz-transition: all 0.4s;
+  transition: all 0.4s;
+
+      &_background-gradient {
+            background: -webkit-linear-gradient(left, #0882FF, #ee7e25, #0882FF);
+      }
+
+}
