+++ packages/client/src/app/pages/Profile/components/ProfileDataList/ProfileDataList.module.scss
@@ -1,0 +1,5 @@
+@import "src/assets/styles/variables";
+
+.form {
+
+}
